+++ Next2/src/Next2/Models/Bindables/SimpleOrderBindableModel.cs
@@ -23,7 +23,7 @@ namespace Next2.Models.Bindables
 
         public int TableNumber { get; set; }
 
-        public double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
 
         public List<SeatModelDTO>? Seats { get; set; }
     }
